add nvtx equivalent for rocm
