+++ CMakeLists.txt
@@ -218,6 +218,9 @@ if(SIRIUS_USE_ROCM)
   if (NOT DEFINED CMAKE_HIP_ARCHITECTURES)
     set(CMAKE_HIP_ARCHITECTURES gfx801 gfx900 gfx90a)
   endif()
+  if(SIRIUS_USE_NVTX)
+    find_package(RocTX REQUIRED)
+  endif()
 
   enable_language(HIP)
   find_package(rocblas CONFIG REQUIRED)
