+++ src/CMakeLists.txt
@@ -125,6 +125,7 @@ target_link_libraries(sirius_cxx PUBLIC ${GSL_LIBRARY}
   $<$<BOOL:${SIRIUS_USE_PUGIXML}>:pugixml::pugixml>
   $<$<BOOL:${SIRIUS_USE_MEMORY_POOL}>:umpire>
   $<$<BOOL:${SIRIUS_USE_NVTX}>:CUDA::nvtx3>
+  $<TARGET_NAME_IF_EXISTS:sirius::roctx>
   $<TARGET_NAME_IF_EXISTS:sirius::cudalibs>
   $<$<BOOL:${SIRIUS_USE_ROCM}>:roc::rocsolver>
   $<$<BOOL:${SIRIUS_USE_ROCM}>:roc::rocblas>
@@ -144,7 +145,7 @@ target_compile_definitions(sirius_cxx PUBLIC
   $<$<BOOL:${SIRIUS_USE_ELPA}>:SIRIUS_ELPA>
   $<$<BOOL:${SIRIUS_USE_NLCGLIB}>:SIRIUS_NLCGLIB>
   $<$<BOOL:${SIRIUS_USE_CUDA}>:SIRIUS_GPU SIRIUS_CUDA>
-  $<$<BOOL:${SIRIUS_USE_NVTX}>:SIRIUS_CUDA_NVTX>
+  $<$<BOOL:${SIRIUS_USE_NVTX}>:SIRIUS_TX>
   $<$<BOOL:${SIRIUS_USE_MAGMA}>:SIRIUS_MAGMA>
   $<$<BOOL:${SIRIUS_USE_ROCM}>:SIRIUS_GPU SIRIUS_ROCM>
   $<$<BOOL:${SIRIUS_USE_VDWXC}>:SIRIUS_USE_VDWXC>
