+++ src/core/profiler.hpp
@@ -19,40 +19,40 @@
 #include <apex_api.hpp>
 #endif
 #include "core/rt_graph.hpp"
-#if defined(SIRIUS_GPU) && defined(SIRIUS_CUDA_NVTX)
-#include "core/acc/nvtx_profiler.hpp"
+#if defined(SIRIUS_GPU) && defined(SIRIUS_TX)
+#include "core/acc/tx_profiler.hpp"
 #endif
 
 namespace sirius {
 
 extern ::rt_graph::Timer global_rtgraph_timer;
 
-#if defined(SIRIUS_CUDA_NVTX)
-extern acc::nvtxprofiler::Timer global_nvtx_timer;
+#if defined(SIRIUS_TX)
+extern acc::txprofiler::Timer global_tx_timer;
 #endif
 
-// TODO: add calls to apex and cudaNvtx
+// TODO: add calls to apex
 
 #if defined(SIRIUS_PROFILE)
 #define PROFILER_CONCAT_IMPL(x, y) x##y
 #define PROFILER_CONCAT(x, y) PROFILER_CONCAT_IMPL(x, y)
 
-#if defined(SIRIUS_CUDA_NVTX)
+#if defined(SIRIUS_TX)
 #define PROFILE(identifier)                                                                                            \
-    acc::nvtxprofiler::ScopedTiming PROFILER_CONCAT(GeneratedScopedTimer, __COUNTER__)(identifier, global_nvtx_timer); \
+    acc::txprofiler::ScopedTiming PROFILER_CONCAT(GeneratedScopedTimer, __COUNTER__)(identifier, global_tx_timer);     \
     ::rt_graph::ScopedTiming PROFILER_CONCAT(GeneratedScopedTimer, __COUNTER__)(identifier, global_rtgraph_timer);
 #define PROFILE_START(identifier)                                                                                      \
-    global_nvtx_timer.start(identifier);                                                                               \
+    global_tx_timer.start(identifier);                                                                                 \
     global_rtgraph_timer.start(identifier);
 #define PROFILE_STOP(identifier)                                                                                       \
     global_rtgraph_timer.stop(identifier);                                                                             \
-    global_nvtx_timer.stop(identifier);
-#else
+    global_tx_timer.stop(identifier);
+#else /* NVTX and ROCTX are not defined -> just use rt_graph */
 #define PROFILE(identifier)                                                                                            \
     ::rt_graph::ScopedTiming PROFILER_CONCAT(GeneratedScopedTimer, __COUNTER__)(identifier, global_rtgraph_timer);
 #define PROFILE_START(identifier) global_rtgraph_timer.start(identifier);
 #define PROFILE_STOP(identifier) global_rtgraph_timer.stop(identifier);
-#endif
+#endif // SIRIUS_CUDA_NVTX
 
 #else
 #define PROFILE(...)
